pyproject.toml
[tool.poetry]
name = "my_app"
version = "0.1.0"
description = ""
authors = ["Younes Piro <piroyounes148@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
mongoengine = "^0.24.0"
django-mongoengine = "^0.4.6"
dnspython = "^2.2.0"
djongo = "^1.3.6"
pymongo = "3.12.3"
graphene-django = "2.13.0"
requests = "^2.27.1"
numpy = "^1.22.3"
beautifulsoup4 = "^4.10.0"
pandas = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
